feat(search): add content type
