+++ apps/strapi/src/extensions/meilisearch/content-types/index.ts
@@ -1,0 +1,7 @@
+import meilisearchIndex from "./meilisearch-index";
+
+const contentTypes = {
+  "meilisearch-index": meilisearchIndex,
+};
+
+export default contentTypes;
