+++ apps/strapi/src/extensions/meilisearch/content-types/meilisearch-index/index.ts
@@ -1,0 +1,7 @@
+import schema from "./schema.json";
+
+const meilisearchIndex = {
+  schema,
+};
+
+export default meilisearchIndex;
